+++ NEWS.md
@@ -26,6 +26,7 @@ Version 0.6.0
  - implement: config shortcuts
  - implement: handle windows logout message
  - fix: editor's inproject property not correctly setted (and may cause devcpp to crash when close project)
+ - implement: print
 
 Version 0.5.0
  - enhancement: support C++ using type alias;